requirements.txt
websockets>=10.4
python-dotenv
llama-cpp-python
langchain_community
langchain_openai
pywin32
pyaudio
SpeechRecognition
pyttsx3
pydub